feat(sac): add channel routing + nav + component extraction (SAC Evolution S2)
