+++ erp/src/app/(app)/sac/reclameaqui/page.tsx
@@ -1,0 +1,14 @@
+"use client";
+
+import { Globe } from "lucide-react";
+import { TicketListPage } from "../shared/ticket-list-page";
+
+export default function ReclameAquiPage() {
+  return (
+    <TicketListPage
+      channelType="RECLAMEAQUI"
+      channelLabel="Reclame Aqui"
+      channelIcon={Globe}
+    />
+  );
+}
