+++ erp/src/app/(app)/sac/whatsapp/page.tsx
@@ -1,0 +1,14 @@
+"use client";
+
+import { MessageSquare } from "lucide-react";
+import { TicketListPage } from "../shared/ticket-list-page";
+
+export default function WhatsAppPage() {
+  return (
+    <TicketListPage
+      channelType="WHATSAPP"
+      channelLabel="WhatsApp"
+      channelIcon={MessageSquare}
+    />
+  );
+}
